requirements.txt
appnope==0.1.0
astroid==1.6.3
bleach==2.1.1
boto==2.48.0
cachetools==2.0.1
certifi==2017.7.27.1
chardet==3.0.4
cycler==0.10.0
Cython==0.27.1
decorator==4.1.2
dill==0.2.7.1
Django==2.0.3
django-cors-headers==2.2.0
django-storages==1.6.5
entrypoints==0.2.3
future==0.16.0
google-api-core==0.1.1
google-auth==1.2.0
google-cloud-core==0.28.0
google-cloud-storage==1.6.0
google-cloud-translate==1.3.1
google-cloud-vision==0.28.0
google-gax==0.15.15
google-resumable-media==0.3.1
googleapis-common-protos==1.5.3
grpcio==1.7.0
html5lib==1.0b10
idna==2.6
ipykernel==4.6.1
ipython==6.2.1
ipython-genutils==0.2.0
ipywidgets==7.0.2
isort==4.3.4
jedi==0.11.0
Jinja2==2.9.6
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.1.0
jupyter-console==5.2.0
jupyter-core==4.3.0
lazy-object-proxy==1.3.1
MarkupSafe==1.0
matplotlib==2.1.0
mccabe==0.6.1
mistune==0.7.4
nbconvert==5.3.1
nbformat==4.4.0
networkx==2.0
notebook==5.2.0
numpy==1.13.3
olefile==0.44
opencv-python==3.3.0.10
pandocfilters==1.4.2
parso==0.1.0
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.3.0
ply==3.8
prompt-toolkit==1.0.15
protobuf==3.4.0
protobuf3-to-dict==0.1.5
psycopg2==2.7.3.2
ptyprocess==0.5.2
pyasn1==0.3.7
pyasn1-modules==0.1.5
Pygments==2.2.0
pylint==1.8.4
pylint-django==0.9.4
pylint-plugin-utils==0.2.6
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2018.3
PyWavelets==0.5.2
pyzmq==16.0.2
qtconsole==4.3.1
requests==2.18.4
rsa==3.4.2
scikit-image==0.13.1
scikit-learn==0.19.0
scipy==0.19.1
simplegeneric==0.8.1
six==1.11.0
terminado==0.6
tesserocr==2.2.2
testpath==0.3.1
tornado==4.5.2
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.0.4
wrapt==1.10.11
